refactor: rest methods accept headers and a url instead of context and route
